refactor: 자동차 이름에 대한 제약조건은 자동차가 관리하도록 변경
